Struct uni::ffi::CString

pub struct CString {
    // some fields omitted
}

Represents an owned C style string

This type generates compatible C-strings from Rust. It guarantees that there is no null bytes in the string and that the string is ended by a null byte.

Methods

impl CString

fn new<T: Into<Vec<u8>>>(t: T) -> Result<CString, ()>

Creates a new C style string

This will ensure that no null byte is present in the string or return and error if that's the case. It will also add the final null byte.

fn as_bytes(&self) -> &[u8]

Return the raw representation of the string without the trailing null byte

fn as_bytes_with_nul(&self) -> &[u8]

Return the raw representation of the string.

This buffer is guaranteed without intermediate null bytes and includes the trailing null byte.