Module uni::hal

Hardware Abstraction Layer

The goal of this module is to abstract the machine/architecture so that the API is the same no matter the underlying hardware.

For now only Xen (x86/x86_64) is supported. A goal for the future would be to be compatible with the virtio standard as well and support Xen arm

Modules

app
Work with the state of the application

arch

mmu
Generic definitions related to the MMU

net
Network device driver abstraction

xen
Xen backend for Uni.rs

Structs

Console
Generic console wrapper

Functions

console
Unprotected access to the hardware console

local_irq_disable
Disable local interrupt delivery

local_irq_enable
Enable local interrupt delivery

local_irq_restore
Restore a local interrupt delivery state

local_irq_save
Disable local interrupt delivery and return the previous state of interrupt delivery

Type Definitions

HwConsoleType
Abstracts the actual hardware console type