Struct uni::hal::mmu::Vaddr

pub struct Vaddr(_);

A virtual address

Methods

impl Vaddr

fn new(value: Vaddr::Repr) -> Self

Create a new virtual address

fn from_ptr<T>(ptr: *const T) -> Self

Create a new virtual address from a constant pointer

fn from_mut_ptr<T>(ptr: *mut T) -> Self

Create a new virtual address from a mutable pointer

fn l4_offset(&self) -> usize

Get the offset in the 4th layer of MMU configuration tables

Note: This value will be garbage if the table does not exist

fn l3_offset(&self) -> usize

Get the offset in the 3rd layer of MMU configuration tables

Note: This value will be garbage if the table does not exist

fn l2_offset(&self) -> usize

Get the offset in the 2nd layer of MMU configuration tables

fn l1_offset(&self) -> usize

Get the offset in the 1st layer of MMU configuration tables

unsafe fn as_ptr<T>(&self) -> *const T

Get the internal value as a constant rust pointer

This is unsafe as the internal value might not be a valid pointer

unsafe fn as_mut_ptr<T>(&mut self) -> *mut T

Get the internal value as a mutable rust pointer

This is unsafe as the internal value might not be a valid pointer

unsafe fn incr(self, count: usize) -> Self

Increment the virtual address

count is the number of bytes you want to increment the pointed value