Struct uni::hal::xen::grant::Ref

pub struct Ref(_);

A grant reference index inside the grant table

Methods

impl Ref

fn grant_access(&self, domid: u16, mfn: Mfn, readonly: bool)

Grant access to a page to the domain with id domid

Note that the page has to be expressed as a Machine Frame Number (mfn)

fn end_access(&mut self)

End a previously granted access