Struct uni::hal::xen::net::XenNetDevice

pub struct XenNetDevice {
    // some fields omitted
}

A Xen vif (virtual interface)

Methods

impl XenNetDevice

fn new(id: u32, intf: InterfaceWeak) -> Result<Box<Self>, ()>

Creates a new Xen network device with id id.

This interface will be the backend of the interface intf