Struct uni::hal::xen::ring::SharedRing

pub struct SharedRing<Req, Resp> {
    // some fields omitted
}

Xen's shared ring

Req is the type of the request

Resp is the type of the response

Methods

impl<Req, Resp> SharedRing<Req, Resp>

fn new(size: usize) -> Option<SharedRing<Req, Resp>>

Allocate a new shared ring

May return None for two reasons: - size is not aligned on PAGE_SIZE - The system is out of memory

unsafe fn from_ptr(ptr: *mut u8, size: usize) -> SharedRing<Req, Resp>

Create a shared ring from an existing memory area

fn req_prod(&self) -> Idx

Returns the request production index

fn rsp_prod(&self) -> Idx

Returns the response production index

fn req_event(&self) -> Idx

Returns the request event index

fn rsp_event(&self) -> Idx

Returns the response event index

unsafe fn req_prod_set(&mut self, req: Idx)

Set the request production index

This is unsafe as an invalid index can corrupt the ring

unsafe fn rsp_prod_set(&mut self, rsp: Idx)

Set the response production index

This is unsafe as an invalid index can corrupt the ring

unsafe fn req_event_set(&mut self, req: Idx)

Set the request event index

unsafe fn rsp_event_set(&mut self, rsp: Idx)

Set the response event index

fn size(&self) -> usize

Returns the size of the ring

Equivalent to __RING_SIZE

fn grant_access(&mut self, domid: u16) -> Option<GrantRef>

Grant access to the ring to a foreign domain with id domid

fn request_from_index(&mut self, idx: usize) -> &mut Req

Get a reference over the request at position idx

fn response_from_index(&mut self, idx: usize) -> &mut Resp

Get a reference over the response at position idx