Module uni::net

Networking primitives for Uni.rs

Modules

conn
Contains various type and trait definition related to connexion management

defs
Various definitions of types, constants, traint, ... related to network

Structs

Builder
Wrap packet creation

Instance
A network stack

InstanceWeak
A weak reference over a network stack

Interface
A shareable network interface.

InterfaceRaw
The raw network interface

InterfaceWeak
A interface weak reference

MultiConn
Connexion that can receive packets from multiple endpoints.

Packet
A network packet

Stack
Uni.rs network stack

UniConn
Connexion the can receive packet from a single endpoint

V4Configuration
IPv4 configuration of an interface

Traits

Formatter
Used to format a packet at the link or network layer