Module uni::net::conn::filter

Definitions of types and traits related to packet filtering

Two different parameters class are going to be mentioned in this module:

Generic parameter: ether type, protocol id, ...

Specific parameter: hardware address, ip address, ...

Structs

GenericFilter
Filter packets on a generic parameter

SpecificFilter
Filter packets on a specific parameter

Traits

Extractor
Extract generic/specific parameters of type T from a packet or a rule.

GenericFilterTrait
Trait implemented by generic filters (i.e. filters based on generic parameters)

PacketSanitizer
Sanitize an incoming packet.

SpecificCallbacks
Callbacks used by SpecificFilter.

SpecificFilterTrait
Trait implemented by specific filters (i.e. filters based on specific parameters)