Struct uni::net::conn::filter::GenericFilter

pub struct GenericFilter<T, U, E, S> where T: Ord + Clone, U: SpecificFilterTrait<T>, E: Extractor<T>, S: PacketSanitizer {
    // some fields omitted
}

Filter packets on a generic parameter

This is a generic class that can be used by protocols to implement a generic filter.

Here is the explanation of types taken as parameters:

T: type of the generic parameter.

U: type of the filter used to filter on specific parameter.

E: type that allows extraction of the generic parameter from rules and packets.

S: type that allows the protocol to sanitize incoming packets.

Methods

impl<T, U, E, S> GenericFilter<T, U, E, S> where T: Ord + Clone, U: SpecificFilterTrait<T>, E: Extractor<T>, S: PacketSanitizer

fn new() -> Self

Create a new generic filter