Trait uni::net::conn::filter::GenericFilterTrait

pub trait GenericFilterTrait {
    fn insert_multi(&mut self, conn: Arc<MultiConn>, rule: &Rule) -> Result<(), ()>;
    fn rx(&self, pkt: Packet);
    fn rx_multi(&self, pkt: Packet, rule: Rule);
}

Trait implemented by generic filters (i.e. filters based on generic parameters)

Generic filters allow protocols to route packets to the appropriate connexion based on a generic parameter.

It may drop the packet if no matching connexion exists or if the packet is invalid (failed checksum, ...).

Required Methods

fn insert_multi(&mut self, conn: Arc<MultiConn>, rule: &Rule) -> Result<(), ()>

Insert a new multi connexion to the filter based on a rule.

fn rx(&self, pkt: Packet)

Filter and route an incoming packet to a connexion (uni or multi).

fn rx_multi(&self, pkt: Packet, rule: Rule)

Filter and route an incoming packet to a multi connexion