Trait uni::net::conn::filter::SpecificCallbacks

pub trait SpecificCallbacks<T: Ord + Clone> {
    fn has_upper_filter(rule: &Rule) -> bool;
    fn filter_from_generic_parameter(param: T) -> Option<Box<GenericFilterTrait>>;
    fn set_layer_rule(rule: &mut Rule, pkt: &Packet);
}

Callbacks used by SpecificFilter.

FIXME: I feel like this is somehow hacky

Required Methods

fn has_upper_filter(rule: &Rule) -> bool

Determine if a rule has an upper protocol component.

For example if you implement a network protocol this will return true if the rule has a transport layer component.

fn filter_from_generic_parameter(param: T) -> Option<Box<GenericFilterTrait>>

Create a new generic filter that filters the generic parameter param of type T.

fn set_layer_rule(rule: &mut Rule, pkt: &Packet)

Set the component of the rule implemented by this protocol based on information contained in a packet.