Trait uni::net::conn::filter::SpecificFilterTrait

pub trait SpecificFilterTrait<T> {
    fn new(generic_param: T) -> Self;
    fn insert_multi(&mut self, conn: Arc<MultiConn>, rule: &Rule) -> Result<(), ()>;
    fn rx(&self, pkt: Packet);
    fn rx_multi(&self, pkt: Packet, rule: Rule);
}

Trait implemented by specific filters (i.e. filters based on specific parameters)

Specific filters allow protocols to route packets to the appropriate connexion based on a specific parameter

It may drop the packet if no matching connexion exists.

Required Methods

fn new(generic_param: T) -> Self

Create a new specific filter.

The generic parameter taken as parameter is used to associate a specific filter with a generic one.

fn insert_multi(&mut self, conn: Arc<MultiConn>, rule: &Rule) -> Result<(), ()>

Insert a new multi connexion to the filter based on a rule.

fn rx(&self, pkt: Packet)

Filter and route an incoming packet to a connexion (uni or multi).

fn rx_multi(&self, pkt: Packet, rule: Rule)

Filter and route an incoming packet to a multi connexion