Struct uni::net::conn::MultiConn

pub struct MultiConn {
    // some fields omitted
}

Connexion that can receive packets from multiple endpoints.

This is just a placeholder object. In order for it to receive packets it must be added to a Manager.

Methods

impl MultiConn

fn new(parent: InterfaceWeak) -> Self

Create a new multi connexion object.

This should not be used directly. Instead, to create a new multi connexion, use Interface::create_multi().

fn pop_packet(&self) -> (Packet, Rule)

Pop a packet from the connexion.

It returns a tuple that contains the packet and a rule that contains information to reply to the sender.

Note that if no packets are available, this function will block until one is received.

fn rx(&self, pkt: Packet, rule: Rule)

Insert a packet inside the connexion

pkt is the received packet

rule is information about the endpoint that sent the packet

fn tx(&self, data: &[u8], rule: &Rule) -> Result<(), ()>

Send data through the connexion.

Since a multi connexion does not target a specific endpoint you have to pass a rule to tell the connexion the endpoint the packet has to be sent to.

fn tx_packet(&self, builder: PacketBuilder, rule: &Rule) -> Result<(), ()>

Send a packet through the connexion.

It might ease the pain of some data sender to use directly a PacketBuilder and send this directly without using a slice.

Since a multi connexion does not target a specific endpoint you have to pass a rule to tell the connexion the endpoint the packet has to be sent to.