Struct uni::net::Builder

pub struct Builder {
    // some fields omitted
}

Wrap packet creation

This object is convenient to create a new packet. It wraps allocation, buffer management and final format of the packet. The packet yielded by the finalize method can be directly sent out on the network.

Methods

impl Builder

fn new() -> Result<Self, ()>

Create a new packet builder

This does an allocation under the hood, which is why this method can fail.

fn size(&self) -> usize

Returns the current size of the packet

fn as_bytes(&self) -> &[u8]

Get the data contained in the packet as a slice

fn set_link_fmt(&mut self, fmt: Arc<Formatter>)

Set the formatter for the link layer

fn set_net_fmt(&mut self, fmt: Arc<Formatter>)

Set the formatter for the network layer

fn write(&mut self, data: &[u8]) -> Result<(), ()>

Write data inside the packet

fn finalize(self, intf: &Interface) -> Result<Packet, ()>

Generate the packet