Struct uni::net::Instance

pub struct Instance(_);

A network stack

This object represent a shareable network stack. This object stack internally uses an Arc so it can be safely shared.

Methods

impl Instance

fn network_thread(instance: Instance)

Network thread linked to an instance

This function is used to process packet that are received within a network stack instance

fn new() -> Self

Create a new network stack

TODO: This cannot really be used more than once for now.

fn downgrade(&self) -> InstanceWeak

Get a weak reference over the network stack

fn interfaces(&self) -> RwLockReadGuard<Vec<Interface>>

Get the list of registered interfaces within network stack

fn enqueue_rx_packet(&self, packet: Packet) -> bool

Enqueue a packet to be received on the network stack

Returns false if the receive queue is full and therefore the packet was not enqueued.

Note: This is safe to be called from interrupt context. Indeed the receive queue is not resizable so no allocation will be performed by this function