Struct uni::net::Packet

pub struct Packet {
    // some fields omitted
}

A network packet

Methods

impl Packet

unsafe fn new(page: *mut u8, offset: usize, size: usize) -> Self

Creates a new packet from a memory pointer page

offset is the offset from the beginning of the page where the actual network data is

size is the size in bytes of the packet

This method is unsafe because a few requirements are necessary:

page must be allocated with __rust_allocate(PAGE_SIZE, PAGE_SIZE)

page ownership is transferred to this packet (i.e. it will be deallocated on Drop)

fn size(&self) -> usize

Returns the size of the packet

fn offset(&self) -> usize

Returns the offset from page() where the data starts

fn page(&self) -> *const u8

Returns the beginning of the allocated page

fn start(&self) -> *const u8

Returns a pointer that points to the beginning of the network data

fn interface(&self) -> Option<Interface>

Return the interface the packet is linked to

fn set_interface(&mut self, intf: InterfaceWeak)

Set the interface the packet is linked to

fn link_hdr_size(&self) -> usize

Returns the size of the link header

This has to be set by the link layer handler

unsafe fn link_hdr_size_mut(&mut self) -> &mut usize

Returns a mutable reference to the size of the link header

fn net_hdr_size(&self) -> usize

Returns the size of the network header

This has to be set by the network layer handler

unsafe fn net_hdr_size_mut(&mut self) -> &mut usize

Returns a mutable reference to the size of the network header

fn tspt_hdr_size(&self) -> usize

Returns the size of the transport header

This has to be set by the transport layer handler

unsafe fn tspt_hdr_size_mut(&mut self) -> &mut usize

Returns a mutable reference to the size of the transport header

fn payload_size(&self) -> usize

Returns the size of the payload.

fn link_header<T>(&self) -> Option<&T>

Get a reference to the link header.

This will return None if the packet size is smaller than the size of the link header.

fn net_header<T>(&self) -> Option<&T>

Get a reference to the network header.

This will return None if the packet size is smaller than the size of the network header.

fn tspt_header<T>(&self) -> Option<&T>

Get a reference to the transport header.

This will return None if the packet size is smaller than the size of the transport header.

fn payload(&self) -> Option<&[u8]>

Extract the payload from the packet.

This will return None if there is no payload