Module uni::sync::spin

Various spin lock definitions

Structs

InterruptSpinGuard

InterruptSpinLock

RwLock
A reader-writer lock

RwLockReadGuard
A guard to which the protected data can be read

RwLockWriteGuard
A guard to which the protected data can be written

SpinGuard
A guard to which the protected data can be accessed

SpinLock
This type provides MUTual EXclusion based on spinning.